WIP: First working version of checklist generation
Doesn't show errors yet :/ but should work otherwise. Also looks way to
hacky in my oppinion with all the conversions needed.

Also needs cleanup (e.g. of checklitst_controller because it doesn't
need the create method anymore). And we probably want to disable editing
checklists for now (in favor of adding copy-on-write duplication in a
later stage).