ci: allow to debug deployment (on PRs) (#32)
# Description

Sometimes deployment changes need manual testing. In these situations,
we'd like to be able to deploy a release candidate without switching to
it.

The proposed changes to the deployment action allows to (optionally) run
deployment on PRs. In case deployments run from a `pull_request` (rather
than a `push`) event, the release is deployed to
`releases/cklist-{version}rc-{git_hash}` and the activating that release
is skipped.

## Checklist

Guess what, we use checklists too :wink:

- [x] I did a self-review of the proposed changes
- [x] I commented my code, particularly in hard-to-understand areas
- [ ] I added automated tests for new functionality (if applicable): N/A