build|ci: Adding postgres DB support (#31)
# Description

The first (static site) version of cklists were generated without DB
support for simplicity. This PR adds support for postgresql DB that will
be used for user authentication and ultimately to store user's
checklists. Getting there, slowly but surely.

In particular, this PR

- [x] adds postgres support in phoenix app
- [x] enables GHA workflows to run DB tests
- [x] changes to the deploy script (to add migrations)

In parallel, a production DB was added on cklist.org.

## Checklist

Guess what, we use checklists too :wink:

- [x] I did a self-review of the proposed changes
- [x] I commented my code, particularly in hard-to-understand areas
- [x] I added automated tests for new functionality (if applicable)