ci: Allow manual deployments for testing (#35)
# Description

This PR adds support for manually triggering test deployments. This was
previously possible by un-commenting "on pull_request", but that's
cumbersome and developers have to remember to properly comment it again.
"on workflow_dispatch" allows to run the workflow manually when needed
from the GitHub web interface.

## Checklist

Guess what, we use checklists too :wink:

- [x] I did a self-review of the proposed changes
- [x] I commented my code, particularly in hard-to-understand areas
- [ ] I added automated tests for new functionality (if applicable)
  N/A

Co-authored-by: Roman Cattaneo <>